Add a copy of pins_arduino.h from 1.0.5 (variants/standard)
The reason is that we can now add certain SODAQ board specifics such as
"Grove Switched" pin. It's not yet added in this version.

Alas, we then also need a complete copy of cores/standard from Arduino IDE
installation. This is inconvenient, but that's how Arduino IDE was
programmed to do. Now we must be aware of potential incompatibilities
with Arduino IDE versions.